% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makeReactiveTrigger}
\alias{makeReactiveTrigger}
\title{makeReactiveTrigger}
\usage{
makeReactiveTrigger()
}
\value{
a list containing functions \code{depend()} and \code{trigger()}
}
\description{
Manually tie together reactive expressions and observers
}
