% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_311.R
\name{page_311}
\alias{page_311}
\title{Title}
\usage{
page_311(input, output, session, coun_dist, week, open_calls = TRUE,
  current_week, snapshots_db)
}
\arguments{
\item{snapshots_db}{}
}
\description{
Title
}
\examples{

}
