% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_311.R
\name{page_311_ui}
\alias{page_311_ui}
\title{311 page UI}
\usage{
page_311_ui(id, open_calls = TRUE)
}
\arguments{
\item{id}{Unique ID for each module instance}

\item{open_calls}{Show open or closed calls?}
}
\value{
A shiny UI
}
\description{
311 page UI
}
