% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_landing.R
\name{page_landing}
\alias{page_landing}
\title{OEM Server}
\usage{
page_landing(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\description{
OEM Server
}
