% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_oem.R
\name{page_oem}
\alias{page_oem}
\title{OEM Server}
\usage{
page_oem(input, output, session, week, coun_dist, snapshots_db)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{week}{reactive value holding selected week}

\item{coun_dist}{reactive value holding selected council district}

\item{snapshots_db}{The pool object holding database connections}
}
\description{
OEM Server
}
