% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{test_shinyapps_io}
\alias{test_shinyapps_io}
\title{Deploy or remove a shinyapps.io app for testing}
\usage{
test_shinyapps_io(host = Sys.getenv("SNAPSHOTS_DB_HOST"),
  user = Sys.getenv("SNAPSHOTS_DB_USER"),
  password = Sys.getenv("SNAPSHOTS_DB_PW"),
  branchName = Sys.getenv("TRAVIS_BRANCH"),
  token = Sys.getenv("RSCONNECT_TOKEN"),
  secret = Sys.getenv("RSCONNECT_SECRET"), remove = FALSE)
}
\arguments{
\item{host}{database hostname}

\item{user}{database username}

\item{password}{database password}

\item{branchName}{branch being deployed from}

\item{token}{rsconnect token}

\item{secret}{rsconnect secret}

\item{remove}{set \code{TRUE} to delete the testing app}
}
\description{
Deploy or remove a shinyapps.io app for testing
}
