% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_311.R
\name{page_311}
\alias{page_311}
\title{Create module server function}
\usage{
page_311(input, output, session, coun_dist, week, open_calls = TRUE,
  current_week, weeks, snapshots_db)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{coun_dist}{reactive value holding selected council district}

\item{week}{reactive value holding selected week}

\item{open_calls}{Show open or closed calls}

\item{current_week}{Current week}

\item{weeks}{List of weeks}

\item{snapshots_db}{The pool object holding database connections}

\item{weeks}{Data frame of week labels}
}
\description{
Create module server function
}
